[Chore] 코드리뷰 반영 수정(#12)
